Move files to avoid deep folders
