+++ source/Databricks/source/SqlStatementExecution/Formats/ApacheArrowFormat.cs
@@ -16,7 +16,6 @@ using System.Collections.Generic;
 using System.Dynamic;
 using System.IO;
 using Apache.Arrow.Ipc;
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Abstractions;
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Configuration;
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Statement;
